+++ crystals/atom.py
@@ -123,7 +123,9 @@ class Atom(Element):
         super().__init__(element=element)
 
         self.coords_fractional = np.asfarray(coords)
-        self.lattice = lattice or Lattice(np.eye(3))
+        self.lattice = lattice or Lattice(
+            np.eye(3)
+        )  # TODO: how to ensure this is never modified afterwards? Maybe emit a warning on modification?
         self.displacement = np.asfarray(
             displacement if displacement is not None else (0, 0, 0)
         )